fix: finalize locale parity review fixes (#1889)

+++ src/Humanizer/Localisation/NumberToWords/OrdinalPrefixScaleNumberToWordsConverter.cs
@@ -190,7 +190,7 @@ class OrdinalPrefixScaleNumberToWordsConverter(OrdinalPrefixScaleNumberToWordsPr
 
         if (hundreds != 0)
         {
-            AddUnit(builder, (ulong)hundreds, GrammaticalGender.Neuter);
+            AddUnit(builder, hundreds, GrammaticalGender.Neuter);
             builder.Add(hundreds == 1 ? profile.HundredSingular : profile.HundredPlural);
         }
